build: handle two defconfig [DO NOT MERGE]
Add the an option to handle two defconfig and the ability to switch between them in the Menu. This allows for faster reconfiguration.

Quick Build (or anything that calls defconfig) is disabled for sanity reasons. Quick Build will totally break the two defconfig option.

The code is extremely messy and untidy, DO NOT MERGE unless you're absolutely sure, it is untested and should not be used.