build: add a way to properly handle older configurations
Newer configurations will have options that older configurations don't, this leads to a whole lot of problems when the script tries to call for the nonexistent new configurations.

Handle this by versioning the configurations, older configurations are marked as "0", if configuration doesn't match the expected valid version, also mark it for reconfiguration.

This fixes our problem with having to delete the configuration folder just for the configuration generator to appear.