install: allow deletion of preexisting dynbw.sh
Let's assume there is a dynbw.sh in install directory, with all possibility, the only reason it would be there is either because of user has installed it once or user copied it there manually.

So, we're safe to delete it since it would be replaced by whatever version exists currently.