dynbw: Cache initial configuration to a temporary file instead
We have utilised input-checking in our codes, but they prompts the user to re-run "config -i". Of course, "config -i" will fail because the script have created the configuration file before.

To fix this, we cache the configuration to a temporary file, which then would be renamed to the configuration file *if* the setup finishes. If not, just trash the temporary file and exit.

* Also add some guard against empty input in toolchain path input field