dynbw: Ignore existing folder if they are not a Git repository
If for some reason the user decides to create a folder that matches the name of the Git repository being sync, they'd encounter an error from Git because the folders usually don't have the Git repository initialised properly.

Ignore these folder from sync because they can cause the sync to fail. The user can still work around this by making an empty git repository or making an empty folder with the name ".git", but at that point, it's more of a problem from the user than from the script itself.