dynbw: Define sync directory in sync.txt instead
Using argument to define our save directory just seems a bit inflexible and odd. We already have the modular sync.txt we can use to define sync directory, so why not define it there?

This removes the need to set directory *all* the time, as sync will pick it up from sync.txt. As a downside, however, our messy code now turns even closer to a spaghetti that is in serious need of a rewrite. I may need some time and actually rewrite this whole thing.

*TODO: Update README to reflect the changes