dynbw: Fix inaccessible help text for build function
Because we call for import_var early on, before even the case statement can process the arguments given, build will simply return warnings and errors if the user doesn't have any configuration file. This is broken for plenty of reasons, but mainly because there are cases when user wants to check help text before creating configuration file, or if the user forgets to create one and jumps to using build instead.

Let's not blame the user for the stupidity and call import_var generously instead. Call import_var whenever we need it instead of greedily assuming the user *will* build something when calling for build.