dynbw: Remove temp_conf_file if it exists in the current directory
We can't rely on our error return call to remove them. It is possible for the user to interrupt the script midway when the configuration has yet to move or delete the temp_conf_file.

However, we don't need to warn the user about it and thus, just delete it silently if we happen to find it around.